apps/traffic-exporter/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
prometheus-client>=0.17.0
kafka-python>=2.0.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
prometheus-client==0.20.0
